+++ components/electric/lib/electric/ddlx/parse/statement.ex
@@ -1,0 +1,7 @@
+defmodule Electric.DDLX.Parse.Statement do
+  defstruct [:stmt, :tokens, :cmd]
+
+  def command(%__MODULE__{} = stmt) do
+    stmt.cmd
+  end
+end
